system-settings: add syncCall, make GenericEnpointCallImpl extend it
